benchmarks/postgres/requirements.txt
asyncpg==0.20.1
dnspython==2.0.0
eventlet==0.25.2
gevent==20.6.2
greenlet==0.4.16
greenletio
monotonic==1.5
psycogreen==1.0.2
psycopg2-binary==2.8.5
six==1.15.0
zope.event==4.4
zope.interface==5.1.0
